Baseline project
Commits the raw initial project to the repo, using a simple "Main Only" branching strategy while also leaving open an option to scale up to a "Feature Isolation" or "Dev  Isolation" strategy in the future.  We are going to build this app using Unity's HDRP in version 2021.3.25f1 (LTS).  For now, we're only building for PC/Standalone, so we'll focus on supporting a keyboard/mouse UI, but we'll leave the door open for mobile and console ports in the future.